onfido/requirements-dev.txt
httpretty==0.8.10
pytest==2.8.2
PyHamcrest==1.8.5
ipython==4.0.0
